Make a script to generate databases
I absolutely *DO NOT ENDORSE* the way the gen_db() function is written.
That is all Achyuth's design.
HORRENDOUS PIECES OF UTTER GARBAGE THAT GUY COMES UP WITH

Everything before it is what programs should look like, cleanly written
abstractions that separate state and operations, not the mismash
of ad-hoc stateful shit web developers are used to.

_sigh_